bug fixed, where the constructor with position argument of a inherited class of gameobj would not work

+++ Characters/BaseCharacter.h
@@ -12,7 +12,7 @@ class BaseCharacter : public BaseGameObject {
 public:
     BaseCharacter();
 
-    BaseCharacter(int x, int y);
+    BaseCharacter(float x, float y);
 
     ~BaseCharacter();
