+++ Characters/Fighter_Andi.h
@@ -10,7 +10,7 @@
 
 class Fighter_Andi : public BaseCharacter {
 public:
-    Fighter_Andi(int x, int y);
+    Fighter_Andi(float x, float y);
     ~Fighter_Andi();
     void update(float deltaTime);
     void draw();
